Refactor Frigate specific details out of ViewMediaBase.
